chore: rename fusionform to hillaform
